+++ Luzyce.Core/Models/ProductionPlan/UpdateInfoInProductionPlanPosition.cs
@@ -5,7 +5,7 @@ public class UpdateInfoInProductionPlanPosition
     public int Id { get; set; }
     public int Quantity { get; set; }
     public int? NumberOfHours { get; set; }
-    public double? WeightNetto { get; set; }
-    public double? WeightBrutto { get; set; }
+    public decimal? WeightNetto { get; set; }
+    public decimal? WeightBrutto { get; set; }
     public int? QuantityPerChange { get; set; }
 }
